+++ src/main/java/com/jupjup/www/jupjup/chat/entity/Room.java
@@ -1,7 +1,6 @@
 package com.jupjup.www.jupjup.chat.entity;
 
-import com.jupjup.www.jupjup.domain.entity.User;
-import com.jupjup.www.jupjup.domain.entity.giveaway.Giveaway;
+import com.jupjup.www.jupjup.giveaway.entity.Giveaway;
 import jakarta.persistence.*;
 import lombok.Builder;
 import lombok.Getter;
