+++ src/main/java/com/jupjup/www/jupjup/giveaway/dto/GiveawayDetailResponse.java
@@ -1,7 +1,7 @@
-package com.jupjup.www.jupjup.model.dto.giveaway;
+package com.jupjup.www.jupjup.giveaway.dto;
 
-import com.jupjup.www.jupjup.domain.entity.giveaway.Giveaway;
-import com.jupjup.www.jupjup.domain.enums.GiveawayStatus;
+import com.jupjup.www.jupjup.giveaway.entity.Giveaway;
+import com.jupjup.www.jupjup.giveaway.enums.GiveawayStatus;
 import com.jupjup.www.jupjup.image.entity.Image;
 import lombok.Builder;
 import lombok.Getter;
