+++ src/test/java/com/jupjup/www/jupjup/chat/repository/RoomRepositoryCustomImplTest.java
@@ -4,7 +4,7 @@
 //import com.jupjup.www.jupjup.chat.entity.QUserChatRoom;
 //import com.jupjup.www.jupjup.chat.entity.Room;
 //import com.jupjup.www.jupjup.domain.entity.User;
-//import com.jupjup.www.jupjup.domain.entity.giveaway.Giveaway;
+//import com.jupjup.www.jupjup.giveaway.entity.Giveaway;
 //import com.jupjup.www.jupjup.domain.repository.GiveawayRepository;
 //import com.jupjup.www.jupjup.domain.repository.UserRepository;
 //import com.querydsl.jpa.impl.JPAQueryFactory;
